% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFormula.R
\name{buildFormula}
\alias{buildFormula}
\title{Builds SDM formulae}
\usage{
buildFormula(
  form.init,
  species,
  sp.formula = NULL,
  sp.partition = NULL,
  useBRMS
)
}
\arguments{
\item{form.init}{The abiotic part of the formula}

\item{species}{The preys (or predators) of the focal species}

\item{sp.formula}{optional parameter for composite variables. See ?trophicSDM}

\item{sp.partition}{optional parameter to specify groups of species for composite variables. See ?trophicSDM}

\item{useBRMS}{whether brms is used (TRUE if penal = "coeff.signs" and method = "stan_glm).}
}
\description{
Builds the formula of both the abiotic and biotic terms to fit a single species SDM based on the input parameters. The function is called inside the SDMfit function
}
\author{
Giovanni Poggiato and Jérémy Andréoletti
}
