% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulData.R
\docType{data}
\name{G}
\alias{G}
\title{Simulated environemntal covariates G}
\format{
A simulated graph of trophic interactions G
}
\usage{
data(G)
}
\description{
Simulated environemntal covariates G
}
\examples{
data(G)
}
\author{
Giovanni Poggiato
}
\keyword{datasets}
