% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SDMfit.R
\name{plot.SDMfit}
\alias{plot.SDMfit}
\title{Plots the regression coeffiecients of a local model}
\usage{
\method{plot}{SDMfit}(x, level = 0.95, plot = TRUE, ...)
}
\arguments{
\item{x}{A SDMfit object, typically obtained with trophicSDM() and available in the field $model of a trophicSDMfit object}

\item{level}{the confidence level of the confidence intervals}

\item{plot}{wheter to print the plot or just return a ggplot object}

\item{...}{additional arguments}
}
\description{
Plots the regression coeffiecients of a local SDMfit model
}
\examples{
data(Y, X, G)
# define abiotic part of the model
env.formula = "~ X_1 + X_2"
# Run the model with bottom-up control using stan_glm as fitting method and no penalisation
# (set iter = 1000 to obtain reliable results)
m = trophicSDM(Y, X, G, env.formula, 
               family = binomial(link = "logit"), penal = NULL, 
               mode = "prey", method = "stan_glm")
# Plot species Y6
plot(m$model$Y6)
}
\author{
Giovanni Poggiato
}
