% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webSDM-package.R
\docType{package}
\name{webSDM-package}
\alias{webSDM}
\alias{webSDM-package}
\title{webSDM: Including Known Interactions in Species Distribution Models}
\description{
A collection of tools to fit and work with trophic Species Distribution Models. Trophic Species Distribution Models combine knowledge of trophic interactions with Bayesian structural equation models that model each species as a function of its prey (or predators) and environmental conditions. It exploits the topological ordering of the known trophic interaction network to predict species distribution in space and/or time, where the prey (or predator) distribution is unavailable. The method implemented by the package is described in Poggiato, Andréoletti, Pollock and Thuiller (2022) \doi{10.22541/au.166853394.45823739/v1}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giopogg/webSDM}
  \item \url{https://giopogg.github.io/webSDM/}
  \item Report bugs at \url{https://github.com/giopogg/webSDM/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Poggiato \email{giov.poggiato@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Jérémy Andréoletti
}

}
\keyword{internal}
